+++ addOns/graphql/CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.18.0] - 2023-07-11
 ### Changed
 - Update minimum ZAP version to 2.13.0.
 - The "Import a GraphQL Schema from a File" and "Import a GraphQL schema from a URL" menu items were merged into one,
@@ -143,6 +143,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   - Import a GraphQL Schema
   - Generate Queries from an imported Schema
 
+[0.18.0]: https://github.com/zaproxy/zap-extensions/releases/graphql-v0.18.0
 [0.17.0]: https://github.com/zaproxy/zap-extensions/releases/graphql-v0.17.0
 [0.16.0]: https://github.com/zaproxy/zap-extensions/releases/graphql-v0.16.0
 [0.15.0]: https://github.com/zaproxy/zap-extensions/releases/graphql-v0.15.0